Upgrade CKEditor to 4.25.0-lts
Upgraded CKEditor from version 4.16.0 to 4.25.0-lts in the file
flaskr/templates/blog/update.html. This change addresses a security
vulnerability present in the older version of CKEditor. By upgrading
to the latest long-term support version, we ensure enhanced security
and stability for the application. The update follows the suggestion
provided in the pull request review comment.