Add Spanish translation support
This commit introduces basic support for Spanish translations on
the main blog page. It allows users to change the language based
on a URL parameter 'locale'. The translations are loaded before
each request using a new 'load_locale' function in the Flask app.

The 'index' function in 'blog.py' now retrieves translations
based on the 'locale' parameter and passes them to the template.
The HTML template 'index.html' has been updated to use the
translation function '_' for translatable strings.

These changes address the need for multilingual support as
described in the project documentation.

Issue ID: FLAS-34