Add Spanish translation support
Implemented basic support for translations on the main blog page.
Users can now change the language based on a URL parameter 'locale'.
The default locale is set to 'en_GB'. Added a before_request
function to load the locale and store it in the global object 'g'.

Updated the blog index view to fetch translations for the current
locale and pass them to the template. Modified the index.html
template to use translation keys for 'Posts', 'New', and 'Edit'.

This change allows the application to support multiple languages,
starting with Spanish, enhancing user accessibility and experience.

Fixes #FLAS-34