Remove pre-commit checks
This commit removes the pre-commit checks from the repository.
The pre-commit configuration files and GitHub workflow
related to pre-commit have been deleted. This change was
made to streamline the development process and reduce
unnecessary checks that were not adding significant value
to the project. By removing these checks, we aim to
simplify the workflow and focus on more critical aspects
of the code review process.

Fixes #FLAS-58