Add markdown support to post body
Implemented markdown support in the post body as per the
specifications outlined in the September 16th demo. This
involves converting markdown to HTML for each post body in
the blog index view. The markdown library is used to achieve
this transformation.

Additionally, updated the create and update templates to
inform users that markdown is supported in the post body.
The index template was also modified to render the post body
safely, allowing HTML content generated from markdown to be
displayed correctly.

Issue ID: FLAS-59