Add markdown support to post body
Implemented markdown support in the post body to allow rendering
of markdown content in the main blog interface. This change
involves converting the post body to HTML using the markdown
library before storing it in the database.

Updated the create and update views to process the post body
with markdown. Modified the templates to render the processed
HTML safely. Added the markdown library to the project
dependencies.

Also, updated tests to verify that markdown content is correctly
rendered as HTML in the post body.

Fixes #FLAS-94